// src/lib/game-data/silk-data/player-data/enum/saveslotcompletionicons_completionstate-enum-silk.generated.ts
/**
 * This file is auto-generated by a script (See ./scripts/*). Do not edit directly.
 */
// This file is generated by scripts/silk-player-data-gen-types.mts. Do not edit manually.

import * as v from 'valibot';

export type SaveSlotCompletionIcons_CompletionStateSilk =
	| 'None' /* 0 */
	| 'Act2Regular' /* 1 */
	| 'Act2Cursed' /* 2 */
	| 'Act2SoulSnare' /* 4 */
	| 'Act3Ending' /* 8 */;
const SaveSlotCompletionIcons_CompletionStateByName: Record<SaveSlotCompletionIcons_CompletionStateSilk, number> = {
	None: 0,
	Act2Regular: 1,
	Act2Cursed: 2,
	Act2SoulSnare: 4,
	Act3Ending: 8,
};
export const SaveSlotCompletionIcons_CompletionStateSilk = {
	isFlags: true,
	byName: SaveSlotCompletionIcons_CompletionStateByName,
	byId: Object.fromEntries(
		Object.entries(SaveSlotCompletionIcons_CompletionStateByName).map(([name, id]) => [id, name] as const),
	) as Record<number, SaveSlotCompletionIcons_CompletionStateSilk>,
	nameList: Object.keys(
		SaveSlotCompletionIcons_CompletionStateByName,
	) as SaveSlotCompletionIcons_CompletionStateSilk[],

	hasFlag(value: number, flag: SaveSlotCompletionIcons_CompletionStateSilk) {
		return (value & SaveSlotCompletionIcons_CompletionStateSilk.byName[flag]) !== 0;
	},
} as const;
export const SaveSlotCompletionIcons_CompletionStateNumberSchemaSilk = v.pipe(
	v.number(),
	v.integer(),
	v.minValue(0),
	v.maxValue(15),
);